Extract thos prepareXXX functions into a new testsPrepreUtils class; refactor SparseTokensFieldTests and BatchClusteringTaskTests
