+++ src/test/java/org/opensearch/neuralsearch/sparse/SparseTokensFieldTests.java
@@ -6,11 +6,6 @@ package org.opensearch.neuralsearch.sparse;
 
 import org.apache.lucene.index.FieldInfo;
 import org.apache.lucene.index.IndexableFieldType;
-import org.apache.lucene.index.IndexOptions;
-import org.apache.lucene.index.DocValuesType;
-import org.apache.lucene.index.DocValuesSkipIndexType;
-import org.apache.lucene.index.VectorEncoding;
-import org.apache.lucene.index.VectorSimilarityFunction;
 import org.junit.Before;
 import org.mockito.MockitoAnnotations;
 
@@ -18,7 +13,7 @@ import java.lang.reflect.Field;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.opensearch.neuralsearch.sparse.algorithm.BatchClusteringTaskTests;
+import org.opensearch.neuralsearch.testsPrepareUtils;
 
 import static org.opensearch.neuralsearch.sparse.SparseTokensField.SPARSE_FIELD;
 
@@ -35,96 +30,7 @@ public class SparseTokensFieldTests extends AbstractSparseTestBase {
 
         fieldName = "testField";
         testValue = new byte[] { 1, 2, 3 };
-        mockType = prepareMockIndexableFieldType();
-    }
-
-    public static IndexableFieldType prepareMockIndexableFieldType() {
-        return new IndexableFieldType() {
-            @Override
-            public boolean stored() {
-                return false;
-            }
-
-            @Override
-            public boolean tokenized() {
-                return false;
-            }
-
-            @Override
-            public boolean storeTermVectors() {
-                return false;
-            }
-
-            @Override
-            public boolean storeTermVectorOffsets() {
-                return false;
-            }
-
-            @Override
-            public boolean storeTermVectorPositions() {
-                return false;
-            }
-
-            @Override
-            public boolean storeTermVectorPayloads() {
-                return false;
-            }
-
-            @Override
-            public boolean omitNorms() {
-                return false;
-            }
-
-            @Override
-            public IndexOptions indexOptions() {
-                return IndexOptions.DOCS_AND_FREQS;
-            }
-
-            @Override
-            public DocValuesType docValuesType() {
-                return DocValuesType.NUMERIC;
-            }
-
-            @Override
-            public DocValuesSkipIndexType docValuesSkipIndexType() {
-                return DocValuesSkipIndexType.NONE;
-            }
-
-            @Override
-            public Map<String, String> getAttributes() {
-                return new HashMap<>();
-            }
-
-            @Override
-            public int pointDimensionCount() {
-                return 0;
-            }
-
-            @Override
-            public int pointIndexDimensionCount() {
-                return 0;
-            }
-
-            @Override
-            public int pointNumBytes() {
-                return 0;
-            }
-
-            @Override
-            public int vectorDimension() {
-                return 0;
-            }
-
-            @Override
-            public VectorEncoding vectorEncoding() {
-                return VectorEncoding.FLOAT32;
-            }
-
-            @Override
-            public VectorSimilarityFunction vectorSimilarityFunction() {
-                return VectorSimilarityFunction.EUCLIDEAN;
-            }
-        };
+        mockType = testsPrepareUtils.prepareMockIndexableFieldType();
     }
 
     public void testSparseTokensFieldConstructor() {
@@ -142,7 +48,7 @@ public class SparseTokensFieldTests extends AbstractSparseTestBase {
     }
 
     public void testIsSparseFieldWhenFieldContainsSparseAttribute() throws Exception {
-        FieldInfo mockField = BatchClusteringTaskTests.prepareKeyFieldInfo();
+        FieldInfo mockField = testsPrepareUtils.prepareKeyFieldInfo("testField");
         Map<String, String> attributes = new HashMap<>();
         attributes.put(SPARSE_FIELD, "true");
